db-dump/pyproject.toml
[tool.poetry]
name = "records_of_processing_activities_db_dump"
version = "0.1.0"
description = "Simple ETL and notebook project to analyse data."
authors = ["Stig B. Dørmænen <stigbd@gmail.com>"]
license = "Apache-2.0"

[tool.poetry.dependencies]
python = "^3.9"
pandas = "^1.2.4"
jupyter = "^1.0.0"
ipykernel = "^5.5.5"

[tool.poetry.dev-dependencies]
pytest = "^6.2.4"
pytest-dotenv = "^0.5.2"
pytest-docker = "^0.10.1"
pytest-cov = "^2.12.0"
coverage = {extras = ["toml"], version = "^5.5"}
black = "^21.5b1"
flake8 = "^3.9.2"
flake8-bandit = "^2.1.2"
flake8-black = "^0.2.1"
flake8-bugbear = "^21.4.3"
flake8-import-order = "^0.18.1"
safety = "^1.10.3"
mypy = "^0.812"
pytype = {version = "^2020.10.08", python = "3.8"}
flake8-annotations = "^2.6.2"
flake8-docstrings = "^1.5.0"
darglint = "^1.8.0"
Pygments = "^2.6.1"
xdoctest = {extras = ["all"], version = "^0.15.4"}
sphinx-autodoc-typehints = "^1.12.0"
codecov = "^2.1.11"
flake8-assertive = "^1.3.0"
Sphinx = "^4.0.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
